Enable basic cookie-based login/logout
As the astute reader will no doubt notice, creating a user profile, or
logging in, requires no password. You can just claim a user name and log
in as that user.

This is obviously insecure, but we'll make do with that for demo
purposes, since the purpose of this code base isn't to enable secure
authentication, but rather to demonstrate use of databases.